backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
dbus-python==1.2.18
exceptiongroup==1.3.1
fastapi==0.128.8
gpg==1.15.1
h11==0.16.0
idna==3.11
libcomps==0.1.18
mysql-connector-python==9.4.0
nftables==0.1
pydantic==2.12.5
pydantic_core==2.41.5
PyGObject==3.40.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
rpm==4.16.1.3
selinux==3.6
six==1.15.0
starlette==0.49.3
systemd-python==234
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
